requirements.txt
asgiref==3.11.0
Django==5.2.9
django-jazzmin==3.0.1
et_xmlfile==2.0.0
openpyxl==3.1.5
polib==1.2.0
sqlparse==0.5.4
tzdata==2025.3
python-dotenv==1.0.1
gunicorn==21.2.0
requests==2.31.0
twilio>=8.0.0
openai>=1.0.0
psycopg2-binary>=2.9.0
dj-database-url>=2.0.0
whitenoise>=6.6.0
